requirements.txt
jsonschema==4.23.0
PyYAML==6.0.2
mypy==1.11.2
ruff==0.6.5
pytest==8.3.2
